requirements.txt
pandas>=1.5.3
scikit_learn>=1.2.2
requests>=2.28.2
Pillow>=9.4.0
reportlab>=4.0.4
matplotlib>=3.7.1
PyPDF2>=3.0.1